Isolate codegen tests from process state
